lax_numpy: factor out indexing tricks
